process: add initial process support

+++ src/arch/aarch64/memory.zig
@@ -1,2 +1,2 @@
-pub const PhsycialAddress: type = usize;
+pub const PhsycialAddress: type = u64;
 pub const VirtualAddress: type = usize;
